requirements.txt
matplotlib==3.9.0
pandas==2.2.2
numpy==1.26.4
Pillow==10.4.0
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
